requirements.txt
fastapi==0.123.4
uvicorn==0.38.0
python-dotenv==1.0.0
PyPDF2==3.0.1
openai==2.8.1
faiss-cpu==1.13.0
python-multipart==0.0.20
pydantic==2.12.5
